Added the ability to build in a docker container
The docker container is based on alpine:3.19 (current latest) and will
automatically download the minimal requirements to build the document.

Texlive "basic" will be installed as a minimal texlive install.
All other required packages are listed in deps.txt.
A user is required to add all new packages to this file for them to be
installed in the container.