Added general minted support
Added the ability to configure certain features of the template within
`misc/setup.tex`. (see line 15 ff.)
This is (for now) only used for the minted support as enabling `minted` is
mutually exclusive with `listings`.
Hence all configuration of the listings is now behind a conditional
checking said configuration flags. (see lines 461 through 517 for the
`listings` configuration and 517 to 525 for `minted`.)

In accordance with the theme of this branch I also made all necessary
changes to the docker container, namely installing `minted`, it's LaTeX
package dependencies and the required python installation.

Lastly the Makefile was also adapted, since using `minted` requires the
additional compiler flag `-shell-escape`.

It has to be noted that currently the listings in chapter 2 are NOT
compatible with minted so a compilation of the whole document with `minted`
enabled will fail.